move exception-related code into pkg/model/error.go
Moves the two exception-processing methods of CRD into
pkg/model/error.go and moves ErrNilShapePointer into pkg/generate which
is the only place it was referenced.
