+++ pkg/model/crd.go
@@ -14,7 +14,6 @@
 package model
 
 import (
-	"errors"
 	"fmt"
 	"sort"
 	"strings"
@@ -380,72 +379,6 @@ func (r *CRD) UpdateConditionsCustomMethodName() string {
 		return ""
 	}
 	return resGenConfig.UpdateConditionsCustomMethodName
-}
-
-// TerminalExceptionCodes returns terminal exception codes as
-// []string for custom resource, if specified in generator config
-func (r *CRD) TerminalExceptionCodes() []string {
-	if r.cfg == nil {
-		return nil
-	}
-	resGenConfig, found := r.cfg.Resources[r.Names.Original]
-	if found && resGenConfig.Exceptions != nil {
-		return resGenConfig.Exceptions.TerminalCodes
-	}
-	return nil
-}
-
-// ExceptionCode returns the name of the resource's Exception code for the
-// Exception having the exception code. If the generator config has
-// instructions for overriding the name of an exception code for a resource for
-// a particular HTTP status code, we return that, otherwise we look through the
-// API model definitions looking for a match
-func (r *CRD) ExceptionCode(httpStatusCode int) string {
-	if r.cfg != nil {
-		resGenConfig, found := r.cfg.Resources[r.Names.Original]
-		if found && resGenConfig.Exceptions != nil {
-			if excConfig, present := resGenConfig.Exceptions.Errors[httpStatusCode]; present {
-				return excConfig.Code
-			}
-		}
-	}
-	if r.Ops.ReadOne != nil {
-		op := r.Ops.ReadOne
-		for _, errShapeRef := range op.ErrorRefs {
-			if errShapeRef.Shape.ErrorInfo.HTTPStatusCode == httpStatusCode {
-				code := errShapeRef.Shape.ErrorInfo.Code
-				if code != "" {
-					return code
-				}
-				return errShapeRef.Shape.ShapeName
-			}
-		}
-	}
-	if r.Ops.ReadMany != nil {
-		op := r.Ops.ReadMany
-		for _, errShapeRef := range op.ErrorRefs {
-			if errShapeRef.Shape.ErrorInfo.HTTPStatusCode == httpStatusCode {
-				code := errShapeRef.Shape.ErrorInfo.Code
-				if code != "" {
-					return code
-				}
-				return errShapeRef.Shape.ShapeName
-			}
-		}
-	}
-	if r.Ops.GetAttributes != nil {
-		op := r.Ops.GetAttributes
-		for _, errShapeRef := range op.ErrorRefs {
-			if errShapeRef.Shape.ErrorInfo.HTTPStatusCode == httpStatusCode {
-				code := errShapeRef.Shape.ErrorInfo.Code
-				if code != "" {
-					return code
-				}
-				return errShapeRef.Shape.ShapeName
-			}
-		}
-	}
-	return "UNKNOWN"
 }
 
 // GoCodeSetExceptionMessagePrefixCheck returns Go code that contains a
@@ -2559,7 +2492,3 @@ func NewCRD(
 		StatusFields:             map[string]*Field{},
 	}
 }
-
-var (
-	ErrNilShapePointer = errors.New("found nil Shape pointer")
-)
